Module forest_cli::cli

Modules

attach_cmd ğŸ”’

auth_cmd ğŸ”’

chain_cmd ğŸ”’

config_cmd ğŸ”’

db_cmd ğŸ”’

fetch_params_cmd ğŸ”’

mpool_cmd ğŸ”’

net_cmd ğŸ”’

send_cmd

shutdown_cmd ğŸ”’

snapshot_cmd ğŸ”’

state_cmd ğŸ”’

sync_cmd ğŸ”’

wallet_cmd ğŸ”’

Structs

Cli

CLI structure generated when interacting with Forest binary

Config ğŸ”’

Enums

Subcommand

Forest binary sub-commands available.

Functions

cli_error_and_die

Print an error message and exit the program with an error code Used for handling high level errors such as invalid parameters

format_vec_pretty ğŸ”’

Format a vector to a prettified string

handle_rpc_err ğŸ”’

Pretty-print a JSON-RPC error and exit

print_rpc_res ğŸ”’

Prints a plain HTTP JSON-RPC response result

print_rpc_res_bytes ğŸ”’

Prints a bytes HTTP JSON-RPC response result

print_rpc_res_cids ğŸ”’

Prints a tipset from a HTTP JSON-RPC response result

print_rpc_res_pretty ğŸ”’

Prints a pretty HTTP JSON-RPC response result

print_stdout ğŸ”’

Prints a string HTTP JSON-RPC response result to a buffered stdout

prompt_confirm ğŸ”’